修改方法返回类型为 Long，以统一保存和删除操作的返回值类型

+++ shengde/src/main/java/icu/nothingless/controller/LoginServlet.java
@@ -36,7 +36,7 @@ public class LoginServlet extends HttpServlet {
         bean.setLastLoginTime(text);
         bean.setRoleId("Admin");
 
-        int t = bean.save();
+        Long t = bean.save();
         System.err.println("Save result: " + t);
         ViewUtil.render(req, resp, "test", Map.of("token", token));
